# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "wtc query API",
    "description": "Versioned read API served by `wtc serve`. The portal SPA generates its typed client from this document. Ingest endpoints (/ingest/*) are source-authenticated (HMAC / bearer / shared-secret header) and out of scope for the portal client.",
    "version": "v1",
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "servers": [{ "url": "/", "description": "The wtc server, relative to its own origin." }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "listEvents",
        "summary": "List change events, newest first, cursor-paginated.",
        "parameters": [
          { "name": "source", "in": "query", "description": "Facet OR-set (comma-separated): github/gitlab/flux/argocd/… or mapping-webhook names.", "schema": { "type": "string" } },
          { "name": "env", "in": "query", "description": "Facet OR-set (comma-separated).", "schema": { "type": "string" } },
          { "name": "service", "in": "query", "description": "Facet OR-set (comma-separated).", "schema": { "type": "string" } },
          { "name": "repo", "in": "query", "description": "Facet OR-set (comma-separated): source repo (owner/name).", "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "description": "Facet OR-set (comma-separated).", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "description": "Facet OR-set (comma-separated).", "schema": { "type": "string" } },
          { "name": "actor", "in": "query", "description": "Facet OR-set (comma-separated).", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "description": "Full-text search over event text.", "schema": { "type": "string" } },
          { "name": "since", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "until", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "cursor", "in": "query", "description": "Opaque cursor from a prior response's next_cursor.", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "A page of events.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventsResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/handoff": {
      "get": {
        "operationId": "handoff",
        "summary": "Summarize deploy activity since an instant (shift-handoff view).",
        "parameters": [
          { "name": "since", "in": "query", "description": "Defaults to 7 days ago.", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": { "description": "Activity summary.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HandoffReport" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "A value from the server's auth.api_tokens." }
    },
    "responses": {
      "BadRequest": { "description": "Malformed request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "Missing or invalid bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
      "Event": {
        "type": "object",
        "required": ["id", "ts", "ingested_at", "source", "kind", "status", "env", "cluster", "namespace", "service", "repo", "actor", "ref", "artifact", "title", "url", "dedup_key"],
        "properties": {
          "id": { "type": "string", "description": "ULID." },
          "ts": { "type": "string", "format": "date-time", "description": "Source event time (UTC)." },
          "ingested_at": { "type": "string", "format": "date-time", "description": "When wtc stored it (UTC)." },
          "source": { "type": "string", "enum": ["github", "flux", "argocd", "helm", "terraform", "manual", "generic", "alertmanager"] },
          "kind": { "type": "string", "enum": ["build", "merge", "push", "deploy", "config_change", "infra_change", "rollback", "alert", "manual"] },
          "status": { "type": "string", "enum": ["started", "succeeded", "failed", "degraded", "unknown"] },
          "env": { "type": "string", "description": "Empty string means unmapped (surfaced by doctor)." },
          "cluster": { "type": "string" },
          "namespace": { "type": "string" },
          "service": { "type": "string" },
          "repo": { "type": "string", "description": "Source repo (owner/name); empty for cluster-side events (flux/argo)." },
          "actor": { "type": "string" },
          "ref": { "type": "string", "description": "Git sha / revision." },
          "artifact": { "type": "string", "description": "Primary artifact, e.g. registry/app:tag." },
          "title": { "type": "string" },
          "url": { "type": "string" },
          "duration_ms": { "type": "integer", "format": "int64", "nullable": true },
          "dedup_key": { "type": "string" },
          "payload": { "type": "string", "description": "Redacted raw JSON; may be omitted." },
          "facts": { "type": "string", "description": "Redacted ingest-time rule facts + pre-rules snapshot (JSON, for /explain); absent when not recorded." }
        }
      },
      "EventsResponse": {
        "type": "object",
        "required": ["events"],
        "properties": {
          "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
          "next_cursor": { "type": "string", "description": "Pass as ?cursor= to fetch the next page; absent on the last page." }
        }
      },
...
      "EnvStats": {
        "type": "object",
        "required": ["total", "failed"],
        "properties": {
          "total": { "type": "integer" },
          "failed": { "type": "integer" }
        }
      },
      "ActorCount": {
        "type": "object",
        "required": ["actor", "count"],
        "properties": {
          "actor": { "type": "string" },
          "count": { "type": "integer" }
        }
      },
      "HandoffReport": {
        "type": "object",
        "required": ["since", "deploys_by_env", "infra_changes", "rollbacks", "unmapped"],
        "properties": {
          "since": { "type": "string", "format": "date-time" },
          "deploys_by_env": { "type": "object", "additionalProperties": { "$ref": "#/components/schemas/EnvStats" } },
          "failures": { "type": "array", "items": { "$ref": "#/components/schemas/Event" }, "description": "Failed deploys in window." },
          "infra_changes": { "type": "integer" },
          "rollbacks": { "type": "integer" },
          "unmapped": { "type": "integer" },
          "top_actors": { "type": "array", "items": { "$ref": "#/components/schemas/ActorCount" } },
          "new_services": { "type": "array", "items": { "type": "string" }, "description": "First ever seen inside the window." }
        }
      },
...
    }
  }
}